+++ src/app/core/services/api/role/role-api.service.ts
@@ -3,6 +3,7 @@ import { ApiService } from '../api.service';
 import { RoleFilter } from '../../../../features/roles/interface/role-filter';
 import { RoleCreate } from '../../../../features/roles/interface/role-create';
 import { RoleUpdate } from '../../../../features/roles/interface/role-update';
+import { RequestQueryParams } from '../../../interface/request-query-params';
 
 @Injectable({
   providedIn: 'root'
@@ -11,11 +12,11 @@ export class RoleApiService {
 
 constructor(private api:ApiService) { }
 
-getRolesList(queryParam:any,data:RoleFilter){
+getRolesList(queryParam:RequestQueryParams,data:RoleFilter){
   return this.api.Post('/Role/GetList',queryParam,data);
 };
-getRoleById(queryParam:any){
-  return this.api.Get('/Role/GetById',queryParam);
+getRoleById(RoleId:string){
+  return this.api.Get('/Role/GetById',{roleId:RoleId});
 };
 CreateRole(data :RoleCreate){
   return this.api.Post('/Role/Create',null,data);
@@ -26,13 +27,13 @@ AddPermissionRole(data :RoleCreate){
 UpdateRole(data :RoleUpdate){
   return this.api.Put('/Role/Update',null,data);
 };
-RemovePermission(RoleId :any){
-  return this.api.Delete('/Role/RemovePermission',RoleId);
+RemovePermission(RoleId :string){
+  return this.api.Delete('/Role/RemovePermission',{roleId:RoleId});
 };
-DeleteRoleById(RoleId :any){
-  return this.api.Delete('/Role/DeleteById',RoleId);
+DeleteRoleById(RoleId :string){
+  return this.api.Delete('/Role/DeleteById',{roleId:RoleId});
 };
-DeleteRoleListById(RoleId :any[]){
-  return this.api.Delete('/Role/DeleteListById',null,RoleId);
+DeleteRoleListById(RoleIds :string[]){
+  return this.api.Delete('/Role/DeleteListById',null,RoleIds);
 };
 }
